+++ cypress/integration/dev-playground/verify-cypress-configuration.spec.ts
@@ -3,7 +3,7 @@ describe('Cypress.config()', () => {
     const myConfig = Cypress.config();
 
     expect(myConfig).to.have.property('animationDistanceThreshold', 5);
-    expect(myConfig).to.have.property('defaultCommandTimeout', 20000);
+    expect(myConfig).to.have.property('defaultCommandTimeout', 25000);
     expect(myConfig).to.have.property('requestTimeout', 18000);
     expect(myConfig).to.have.property('responseTimeout', 30000);
     expect(myConfig).to.have.property('pageLoadTimeout', 60000);
