+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/add-hotel-travel-policy.ts
