+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/cancel-delete-dialog-and-confirm.ts
@@ -2,7 +2,7 @@ import { TravelPolicy } from '@cy/models';
 import { clickTravelPolicyCtaButton } from './click-travel-policy-cta-button';
 
 export const cancelDeleteDialogAndConfirm = (travelPolicyDetails: TravelPolicy): void => {
-  clickTravelPolicyCtaButton(travelPolicyDetails.sharedDetails.modifiedName, 'delete');
+  clickTravelPolicyCtaButton(travelPolicyDetails.sharedDetails.modifiedName, 'Delete');
 
   cy.get('.cdk-overlay-container confirmation-dialog ui-button[cancel=true]').click();
