+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/click-travel-policy-cta-button.ts
