+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/close-edit-dialog-and-confirm.ts
@@ -2,7 +2,7 @@ import { TravelPolicy } from '@cy/models';
 import { clickTravelPolicyCtaButton } from './click-travel-policy-cta-button';
 
 export const closeEditDialogAndConfirm = (travelPolicyDetails: TravelPolicy): void => {
-  clickTravelPolicyCtaButton(travelPolicyDetails.sharedDetails.name, 'edit');
+  clickTravelPolicyCtaButton(travelPolicyDetails.sharedDetails.name, 'Edit');
 
   cy.get('.cdk-overlay-container dib-dialog-wrapper i').contains('close').click();
