+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/delete-travel-policy-and-confirm.ts
