+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/delete-travel-policy.ts
@@ -1,7 +1,7 @@
 import { clickTravelPolicyCtaButton } from './click-travel-policy-cta-button';
 
 export const deleteTravelPolicy = (travelPolicyName: string): void => {
-  clickTravelPolicyCtaButton(travelPolicyName, 'delete');
+  clickTravelPolicyCtaButton(travelPolicyName, 'Delete');
 
   cy.get('.cdk-overlay-container confirmation-dialog ui-button[type=warning]').click();
 };
