+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/shared/enter-shared-details.ts
