+++ cypress/integration/dib-travel/company-settings/travel-settings/travel-policy/train.spec.ts
@@ -8,18 +8,18 @@ import {
   searchAndSelectEmployee,
 } from './shared';
 
-describe('Company Settings - Travel Policy - Train', () => {
+describe('Company Settings - Travel Settings - Travel Policy - Train', () => {
   let travelPolicyDetails: TravelPolicy;
 
   before(() => {
-    cy.fixture('company-settings/travel-policy-details').then((travelPolicyDetailsFixture) => {
+    cy.fixture('company-settings/travel-settings-details').then((travelPolicyDetailsFixture) => {
       travelPolicyDetails = travelPolicyDetailsFixture;
     });
   });
 
   beforeEach(() => {
     cy.login();
-    cy.visit('/company-management/travel-policy');
+    cy.visit('/company-management/travel-settings');
   });
 
   it('should add train travel policy', () => {
