+++ cypress/models/company-settings/index.ts
@@ -1,4 +1,4 @@
 export * from './approval-process-group';
 export * from './payment-method';
 export * from './reference-fields';
-export * from './travel-policy';
+export * from './travel-settings';
